requirements.txt
appnope==0.1.3
asgiref==3.7.2
asttokens==2.3.0
backcall==0.2.0
certifi==2023.7.22
charset-normalizer==3.2.0
comm==0.1.4
contourpy==1.1.0
cycler==0.11.0
debugpy==1.6.7.post1
decorator==5.1.1
Django==4.2.5
django-cors-headers==4.2.0
django-picklefield==3.1
djangorestframework==3.14.0
et-xmlfile==1.1.0
executing==1.2.0
filelock==3.12.3
fonttools==4.42.1
idna==3.4
ipykernel==6.25.1
ipython==8.15.0
ipywidgets==8.1.0
jedi==0.19.0
Jinja2==3.1.2
joblib==1.3.2
jupyter_client==8.3.1
jupyter_core==5.3.1
jupyterlab-widgets==3.0.8
kiwisolver==1.4.5
MarkupSafe==2.1.3
matplotlib==3.7.2
matplotlib-inline==0.1.6
mpmath==1.3.0
nest-asyncio==1.5.7
networkx==3.1
numpy==1.25.2
openpyxl==3.1.2
packaging==23.1
pandas==2.1.0
parso==0.8.3
patsy==0.5.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.0
platformdirs==3.10.0
prompt-toolkit==3.0.39
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.16.1
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
pyzmq==25.1.1
requests==2.31.0
scikit-learn==1.3.0
scipy==1.11.2
seaborn==0.12.2
six==1.16.0
sklearn==0.0.post7
sqlparse==0.4.4
stack-data==0.6.2
statsmodels==0.14.0
sympy==1.12
threadpoolctl==3.2.0
torch==2.0.1
torchvision==0.15.2
tornado==6.3.3
tqdm==4.66.1
traitlets==5.9.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
wcwidth==0.2.6
widgetsnbextension==4.0.8
